# services/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateHotelBody:
    properties:
      address:
        example: 1 Chiyoda, Tokyo
        type: string
      description:
        example: Luxury hotel in central Tokyo
        type: string
      location:
        $ref: '#/definitions/dto.LocationDTO'
      owner_id:
        example: 1
        type: integer
      title:
        example: Imperial Hotel Tokyo
        type: string
    type: object
...
  dto.HotelResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      description:
        type: string
      hotel_slug:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/dto.LocationDTO'
      owner_id:
        type: integer
      rating:
        type: number
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.HotelShortResponse:
    properties:
      address:
        type: string
      hotel_slug:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/dto.LocationDTO'
      owner_id:
        type: integer
      rating:
        type: number
      title:
        type: string
    type: object
  dto.HotelsResponse:
    properties:
      hotels:
        items:
          $ref: '#/definitions/dto.HotelShortResponse'
        type: array
    type: object
  dto.LocationDTO:
    properties:
      latitude:
        example: 35.66
        type: number
      longitude:
        example: 139.7
        type: number
    type: object
...
  dto.UpdateHotelRequest:
    properties:
      address:
        example: 2 Chiyoda, Tokyo
        type: string
      description:
        example: Renovated rooms and improved amenities
        type: string
      location:
        $ref: '#/definitions/dto.LocationDTO'
    type: object
  dto.UpdateHotelTitleRequest:
    properties:
      title:
        example: Imperial Hotel Tokyo Annex
        type: string
    type: object
...
  responder.ErrorResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/responder.ErrorResponseDetails'
        type: array
      error:
        type: string
    type: object
  responder.ErrorResponseDetails:
    properties:
      description:
        type: string
      field:
        type: string
    type: object
info:
  contact: {}
  description: REST API Gateway for Fukuro Reserve microservices.
  title: Swagger Gateway API
  version: "1.0"
paths:
  /{countryCode}/{citySlug}/hotels:
    get:
      description: Public endpoint. Returns hotels in location with optional sorting
        and pagination.
      parameters:
      - description: Country code (ISO 3166-1 alpha-2)
        example: jp
        in: path
        name: countryCode
        required: true
        type: string
      - description: City slug
        example: tokyo
        in: path
        name: citySlug
        required: true
        type: string
      - default: title
        description: Sort field
        enum:
        - title
        - rating
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - default: 1
        description: Page number (starts from 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HotelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      summary: Get hotels
      tags:
      - hotels
    post:
      consumes:
      - application/json
      description: Creates a hotel in the selected location. Requires JWT auth and
        role USER_ROLE_MODERATOR or USER_ROLE_ADMIN.
      parameters:
      - description: Country code (ISO 3166-1 alpha-2)
        example: jp
        in: path
        name: countryCode
        required: true
        type: string
      - description: City slug
        example: tokyo
        in: path
        name: citySlug
        required: true
        type: string
      - description: Create hotel request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateHotelBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.HotelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new hotel
      tags:
      - hotels
  /{countryCode}/{citySlug}/hotels/{hotelId}:
    delete:
      description: Deletes hotel by ID. Requires JWT auth.
      parameters:
      - description: Country code (ISO 3166-1 alpha-2)
        example: jp
        in: path
        name: countryCode
        required: true
        type: string
      - description: City slug
        example: tokyo
        in: path
        name: citySlug
        required: true
        type: string
      - description: Hotel ID
        example: 0f8fad5b-d9cb-469f-a165-70867728950e
        in: path
        name: hotelId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete hotel
      tags:
      - hotels
    get:
      description: Public endpoint. Returns hotel details by hotel ID.
      parameters:
      - description: Country code (ISO 3166-1 alpha-2)
        example: jp
        in: path
        name: countryCode
        required: true
        type: string
      - description: City slug
        example: tokyo
        in: path
        name: citySlug
        required: true
        type: string
      - description: Hotel ID
        example: 0f8fad5b-d9cb-469f-a165-70867728950e
        in: path
        name: hotelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HotelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      summary: Get hotel by ID
      tags:
      - hotels
    put:
      consumes:
      - application/json
      description: Fully updates mutable hotel fields. Requires JWT auth.
      parameters:
      - description: Country code (ISO 3166-1 alpha-2)
        example: jp
        in: path
        name: countryCode
        required: true
        type: string
      - description: City slug
        example: tokyo
        in: path
        name: citySlug
        required: true
        type: string
      - description: Hotel ID
        example: 0f8fad5b-d9cb-469f-a165-70867728950e
        in: path
        name: hotelId
        required: true
        type: string
      - description: Update hotel request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateHotelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HotelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Update hotel
      tags:
      - hotels
...
  /{countryCode}/{citySlug}/hotels/{hotelId}/title:
    patch:
      consumes:
      - application/json
      description: Partially updates hotel title and slug. Requires JWT auth.
      parameters:
      - description: Country code (ISO 3166-1 alpha-2)
        example: jp
        in: path
        name: countryCode
        required: true
        type: string
      - description: City slug
        example: tokyo
        in: path
        name: citySlug
        required: true
        type: string
      - description: Hotel ID
        example: 0f8fad5b-d9cb-469f-a165-70867728950e
        in: path
        name: hotelId
        required: true
        type: string
      - description: Update hotel title request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateHotelTitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HotelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Update hotel title
      tags:
      - hotels
  /{countryCode}/{citySlug}/hotels/slug/{hotelSlug}:
    get:
      description: Public endpoint. Returns hotel details by SEO-friendly hotel slug.
      parameters:
      - description: Country code (ISO 3166-1 alpha-2)
        example: jp
        in: path
        name: countryCode
        required: true
        type: string
      - description: City slug
        example: tokyo
        in: path
        name: citySlug
        required: true
        type: string
      - description: Hotel slug
        example: imperial-hotel-tokyo
        in: path
        name: hotelSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HotelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      summary: Get hotel by slug
      tags:
      - hotels
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"